+++ CHANGELOG.rst
@@ -14,5 +14,7 @@ Added
 
 * Minimum viable product (MVP) functionality.
   Updates the screen with running counts.
-* Default to sorting output alphabetically
-  to more closely match the behavior of ``sort | uniq --count``.
+* Imitates the output format of ``sort | uniq --count``
+  by sorting output alphabetically,
+  displaying the count first,
+  and right justifying the count by padding with up to three spaces.